+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#11](https://gitlab.com/tozd/waf/-/issues/11)
 - Use route priority to match Vue Router 5.
   [#13](https://gitlab.com/tozd/waf/-/issues/13)
-- Remove `--tls.email` CLI flag. Now enable Let's Encrypt by setting `--tls.cache`.
+- Rename `--tls.*` CLI flags to `--https.*`.
+  Rename `--tls.cache` CLI flag to `--https.letsencrypt`.
+  Remove `--tls.email` CLI flag. Now enable Let's Encrypt by setting `--https.letsencrypt`.
 
 ## [0.21.0] - 2025-10-13